+++ tests/Type/ExternalReplyInfoTest.php
@@ -183,6 +183,7 @@ final class ExternalReplyInfoTest extends TestCase
                 'type' => 'regular',
                 'allows_multiple_answers' => true,
                 'allows_revoting' => false,
+                'members_only' => true,
             ],
             'venue' => [
                 'location' => [